Make switching to privileged mode step 2.
Relevant thread
--------------------

Message from Murali sir:
I think Alan's point is that unless the machine has switched to unprivileged code, how is there an access to something outside it's address space (IVT).

Technically speaking, what Farzeen told is right  Afterall, during the execution of an un-privileged instruction like MOV or PUSH, there is implicit access to page table in memory as well as registers like PTLR and PTBR which are not accessible to user programs directly.  So, what is wrong with implicit access to memory (vector table)?   

Nevertheless, it is convenient to say that after saving the return address to stack,  the machine switches to privileged mode before accessing IVT.  This can avoid the confusion which Alan had (and some others might also have).  Moreover, after the transfer of IP to stack, there is not much gained by delaying the announcement about entering protected mode.    
 
https://groups.google.com/g/xos-users/c/MqHa9taxgxc/m/jnewPlQJCgAJ